requirements.txt
aiosmtplib==3.0.2
alembic==1.16.5
annotated-types==0.7.0
anyio==4.10.0
appdirs==1.4.4
apturl==0.5.2
asgiref==3.5.0
attrs==21.2.0
backcall==0.2.0
bcrypt==3.2.0
beautifulsoup4==4.10.0
beniget==0.4.1
blinker==1.9.0
Brlapi==0.8.3
Brotli==1.0.9
certifi==2020.6.20
chardet==4.0.0
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==3.4.8
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.2.18
decorator==4.4.2
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
dnspython==2.8.0
duplicity==0.8.21
ecdsa==0.19.1
email-validator==2.3.0
entrypoints==0.4
exceptiongroup==1.3.0
fastapi==0.116.1
fastapi-mail==1.5.0
fasteners==0.14.1
fonttools==4.29.1
fs==2.4.12
future==0.18.2
gast==0.5.2
greenlet==3.2.4
h11==0.16.0
html5lib==1.1
httplib2==0.20.2
httptools==0.6.4
idna==3.3
importlib-metadata==4.6.4
ipykernel==6.7.0
ipython==7.31.1
ipython_genutils==0.2.0
jedi==0.18.0
jeepney==0.7.1
Jinja2==3.1.6
jupyter-client==7.1.2
jupyter-core==4.9.1
keyring==23.5.0
kiwisolver==1.3.2
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
lxml==4.8.0
lz4==3.1.3+dfsg
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
monotonic==1.6
more-itertools==8.10.0
mpmath==0.0.0
nest-asyncio==1.5.4
netifaces==0.11.0
numpy==1.21.5
oauthlib==3.2.0
olefile==0.46
packaging==21.3
paramiko==2.9.3
parso==0.8.1
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
ply==3.11
prompt-toolkit==3.0.28
protobuf==3.12.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
py==1.10.0
pyasn1==0.6.1
pycairo==1.20.1
pycups==2.0.1
pydantic==2.11.9
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.11.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.1.1
python-jose==3.5.0
pythran==0.10.0
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
pyzmq==22.3.0
reportlab==3.6.8
requests==2.25.1
rsa==4.9.1
scipy==1.8.0
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
soupsieve==2.3.1
SQLAlchemy==2.0.43
starlette==0.47.3
sympy==1.9
systemd-python==234
tomli==2.2.1
tornado==6.1
traitlets==5.1.1
typing-inspection==0.4.1
typing_extensions==4.15.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufoLib2==0.13.1
ufw==0.36.1
unattended-upgrades==0.1
unicodedata2==14.0.0
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.35.0
uvloop==0.21.0
wadllib==1.3.6
watchfiles==1.1.0
wcwidth==0.2.5
webencodings==0.5.1
websockets==15.0.1
wsproto==1.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0

# httpx is required by FastAPI TestClient in our local tests
httpx==0.24.1
